changes for logstreaming
